v0.5.3
Start of v0.5.3 development branch. Different enough to get a minor version update. This extends the work of @jmcurran to use the blandr object class. This is a better solution that can now use generic versions of print, summary, plot and ggplot. Hopefully can then tidy up the unholy mess that is the current large list of functions and trim them down